# Improvements to System.Math InternalCall code.