--- a/src/classlibnative/float/Float.nativeproj
+++ b/src/classlibnative/float/Float.nativeproj
@@ -29,6 +29,7 @@
   <!--Leaf Project Items-->
   <ItemGroup>
     <CppCompile Include="floatdouble.cpp" />
+    <CppCompile Include="floatsingle.cpp" />
   </ItemGroup>
   <!--Import the targets-->
   <Import Project="$(_NTDRIVE)$(_NTROOT)\ndp\clr\clr.targets" />