--- a/src/classlibnative/inc/floatdouble.h
+++ b/src/classlibnative/inc/floatdouble.h
@@ -2,8 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#ifndef _FLOATCLASS_H_
-#define _FLOATCLASS_H_
+#ifndef _FLOATDOUBLE_H_
+#define _FLOATDOUBLE_H_
 
 #include <object.h>
 #include <fcall.h>
@@ -41,4 +41,4 @@ public:
 };
     
 
-#endif // _FLOATCLASS_H_
+#endif // _FLOATDOUBLE_H_