--- a/src/pal/tests/palsuite/c_runtime/_finite/test1/testinfo.dat
+++ b/src/pal/tests/palsuite/c_runtime/_finite/test1/testinfo.dat
@@ -9,6 +9,5 @@ Name = Positive Test for _finite
 TYPE = DEFAULT
 EXE1 = test1
 Description
-=Checks that _finite correctly classifies all types of floating 
-=point numbers
-=(NaN, -Infinity, Infinity, finite nonzero, unnormalized, 0, and -0).
+= Checks that _finite correctly classifies all types of floating point
+= numbers (NaN, -Infinity, Infinity, finite nonzero, unnormalized, 0, and -0).