--- a/src/pal/tests/palsuite/c_runtime/_isnan/CMakeLists.txt
+++ b/src/pal/tests/palsuite/c_runtime/_isnan/CMakeLists.txt
@@ -1,4 +1,3 @@
 cmake_minimum_required(VERSION 2.8.12.2)
 
 add_subdirectory(test1)
-