--- a/src/pal/tests/palsuite/c_runtime/asin/test1/testinfo.dat
+++ b/src/pal/tests/palsuite/c_runtime/asin/test1/testinfo.dat
@@ -12,6 +12,3 @@ Description
 = Passes a series of values to the asin() function,
 = checking each for the expected result. Also checks
 = for proper handling of out-of-range values.
-
-
-