--- a/src/pal/tests/palsuite/c_runtime/atan/test1/testinfo.dat
+++ b/src/pal/tests/palsuite/c_runtime/atan/test1/testinfo.dat
@@ -11,6 +11,3 @@ EXE1 = test1
 Description
 = Passes a series of values to the atan() function,
 = checking each for the expected result.
-
-
-