--- a/src/pal/tests/palsuite/c_runtime/cos/test1/testinfo.dat
+++ b/src/pal/tests/palsuite/c_runtime/cos/test1/testinfo.dat
@@ -11,6 +11,3 @@ EXE1 = test1
 Description
 = Passes to cos() a series of angle value, checking that
 = each one return the correct value.
-
-
-