… src/classlibnative/float/CMakeLists.txt
… src/classlibnative/float/floatnative.h
… src/classlibnative/float/floatsingle.cpp
… src/pal/src/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/_isnan/test1/test1.c
… src/pal/tests/palsuite/c_runtime/acos/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/acos/test1/test1.c
… src/pal/tests/palsuite/c_runtime/asin/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/asin/test1/test1.c
… src/pal/tests/palsuite/c_runtime/atan/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/atan/test1/test1.c
… src/pal/tests/palsuite/c_runtime/atan2/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/atan2/test1/test1.c
… src/pal/tests/palsuite/c_runtime/ceil/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/ceil/test1/test1.c
… src/pal/tests/palsuite/c_runtime/cos/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/cos/test1/test1.c
… src/pal/tests/palsuite/c_runtime/cosh/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/cosh/test1/test1.c
… src/pal/tests/palsuite/c_runtime/exp/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/exp/test1/test1.c
… src/pal/tests/palsuite/c_runtime/exp/test2/test2.c
… src/pal/tests/palsuite/c_runtime/exp/test3/test3.c
… src/pal/tests/palsuite/c_runtime/fabs/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/fabs/test1/test1.c
… src/pal/tests/palsuite/c_runtime/fabs/test1/testinfo.dat
… src/pal/tests/palsuite/c_runtime/fabsf/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/fabsf/test1/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/fabsf/test1/test1.c