This cleans up the native code for the `System.Math` internal calls. There were several instances of workarounds implemented for much older versions of the MSVC compiler that are just not relevant anymore.

The list of workarounds removed are as follows:
* On Windows AMD64, `Sin`/`Cos`/`Tan` were implement in `vm\amd64\JitHelpers_Fast.asm` (using x87 floating-point instructions) because the CRT versions were previously too slow.
* `Log`, `Log10`, `Exp`, `Pow`, and `Round` were being compiled with `/fp:precise` because the `/fp:fast` implementation was previously slower.
* `Exp(+/-INFINITY)` was previously special handled as the CRT implementation did not return the expected values.
* On Windows x86, `Pow` had some additional code (and inline assembly) that was unnecessary.
* On Windows x86, `Round` was using inline assembly to call the `frndint` x87 floating-point instruction

In all of these cases, the replaced code was one or more of the following:
* No longer needed as the bug had been fixed
* Went against the Intel/AMD CPU Optimization Guidelines by using x87 floating-point instructions. Both manuals recommend using SSE/SSE2 instructions unless their are special circumstances requiring the use of the x87 floating-point instructions; the SSE/SSE2 instructions can be pipelined, have a lower latency, use hardware registers rather than the floating-point stack, and generally provided faster computations than the special x87 floating point instructions (such as `fsin`, `fcos`, and `ftan`).
* Was no longer more performant than the 'correct' implementation