@tannergooding 2016-05-09
This should be reviewable. Everything is passing and looking good, so I don't have any other changes I am planning on making (unless otherwise requested).