@jkotas 2016-05-10
@tannergooding Could you please split this into several smaller PRs? It will be hard to make progress on it as is because of it has too much stuff in it.
- File rename so that the diff for the rest looks more reasonable
- Non-x86 specific changes
- x86 specific changes. How did you test that these changes are good?
Thanks!