@AndyAyersMS 2016-05-19
@tannergooding any plans to implement perf tests for these math functions?

Also how about more rigorous correctness tests? 

There are a few tests over in CoreFx but they're minimal and don't check boundary cases. There are pal tests for the underlying native methods but nothing that verifies these when called from managed code.