@tannergooding 2016-05-19
@AndyAyersMS. I'll see about getting some perf tests implemented as well as some more rigorous correctness tests.