@AndyAyersMS 2016-05-20
I wrote [a simple perf test](https://github.com/AndyAyersMS/coreclr/blob/MathBenchmarks/tests/src/JIT/Performance/CodeQuality/Math/Functions.cs) for `Log`, feel free to build on that if you like.

BTW math function performance seems significantly worse on Linux than on Windows.