@tannergooding 2016-05-22
I have added some very basic performance tests for all of the `System.Math` functions which are implemented via `InternalCall`. I will be working on additionally adding some more functional tests.

All performance tests are implemented as follows:
* 100,000 iterations are executed
* The time of all iterations are aggregated to compute the `Total Time`
* The time of all iterations are averaged to compute the `Average Time`
* A single iteration executes some simple operation, using the function under test, 5000 times

All of the functions which were not modified did not show any improvement or regression.

The execution time below is the `Total Time` for all 100,000 iterations, measured in seconds.

Of the functions which were changed, they are showing the following improvments (x64):
|Function|Improvment|Execution Time - Before|Execution Time - After|
|:--:|:--:|:--:|:--:|
|Cos  | 59.8373609%|17.5828667s|7.06169908s|
|Exp  | 13.0115399%|7.78959558s|6.77604924s|
|Log  |0.691849597%|5.47524991s|5.43736941s|
|Log10|0.214979727%|5.79203835s|5.77961335s|
|Pow  |0.982046086%|30.8941570s|30.5907621s|
|Round| 31.0027713%|5.90780906s|4.07622453s|
|Sin  | 68.0349993%|17.6330190s|5.63639463s|
|Tan  | 59.4324800%|25.4978910s|10.3438620s|

Of the functions which were changed, they are showing the following improvments (x86):
|Function|Improvment|Perf Before|Perf After|
|:--:|:--:|:--:|:--:|
|Exp  | 21.8584468%|9.46195789s|7.39372086s|
|Log  | 14.4277698%|11.2499700s|9.62685020s|
|Log10| 14.9602930%|11.5140937s|9.79155157s|
|Pow  | 15.1681239%|31.8424078s|27.0125119s|
|Round|0.172920825%|0.63726968s|0.63616771s|

FYI. @AndyAyersMS