@AndyAyersMS 2016-05-22
Thanks for adding tests.... left a few comments.

I'd be interested to see the Windows vs Linux numbers.