@tannergooding 2016-05-22
I've updated according to the feedback provided and am working on getting some Linux (and Mac) numbers for comparison.