@AndyAyersMS 2016-05-22
Thanks for the updates.

One last thing to think about. The code under jit/performance has to work well under 3 different use cases:
* Regular test, run every CI. We do this a lot
* Perf test, xunit-perf run. We do this rarely
* Perf test, manual run. We do this very rarely.

I think what you have is good on the perf test aspects, but it's likely running this as a regular test will take a long time. What I've done for the other perf tests is (for now) give up on the manual perf test aspect and scale back iterations so when run as an exe the timing is reasonable, say a second or so at most in release and a short as possible in non-release.

To get back the manual mode capabilities, I have plans to go back and add optional `-bench` arguments to each of tests so they ramp up the work they do when run as a standalone perf test.

So, look at how long this test takes in CI and if it stands out, you should consider doing something to trim it back.