@AndyAyersMS 2016-05-22
Looks like you hit the timeout ... 
```
D:\j\workspace\debug_windows---17180f1b\bin\tests\Windows_NT.x64.Debug\JIT\Performance\CodeQuality\Math\Functions\Functions\Functions.cmd Timed Out
```