@tannergooding 2016-05-22
Updated. The Math Functions performance test now accepts two arguments. One allows you to specify the name of the test to run (`all` is supported and is the default) and the other (`-bench`) allows you to specify the number of iterations to run (defaults to `1` on debug and `1000` on release).