@tannergooding 2016-05-23
I have some perf numbers for Linux and Mac. I'm not sure I trust the Linux numbers because of how small they are (I did double check the I deployed the right bits), but Mac did show some small (and consistent) improvements (I was happy to see already good perf coming from Mac). I also updated my post above to list the hardware I ran the windows benchmarks on.

OS X showed very little improvement, but the improvement was consistent.

**Hardware: MacBook Pro w/ 2.5GHz Quad-Core i7 (Intel) and 16GB RAM**

|Function|Improvment|Execution Time - Before|Execution Time - After|
|:--:|:--:|:--:|:--:|
|Cos  | 8.38015985%|5.39619623s|4.94398636s|
|Exp  | 3.44344613%|3.70642912s|3.57880023s|
|Log  |0.155915035%|3.37851958s|3.37325196s|
|Log10|0.274460120%|3.40944615s|3.40008858s|
|Pow  | 1.00818607%|15.5424187s|15.3857222s|
|Round|0.243108725%|2.61207820s|2.60572801s|
|Sin  |0.572526337%|4.50698218s|4.48117852s|
|Tan  | 23.0676004%|5.25166822s|4.04023438s|

Linux showed even less improvement and it did not seem to be consistent.

**Hardware: Asus ZenBook w/ 2.6GHz Quad-Core i3 (Intel) and 4GB RAM**

|Function|Improvment|Execution Time - Before|Execution Time - After|
|:--:|:--:|:--:|:--:|
|Cos  | 0.209634293%|54.7168588s|54.6021535s|
|Exp  |  1.89459115%|52.8744791s|51.8727239s|
|Log  | 0.217855270%|56.1605876s|56.0382388s|
|Log10| 0.274460120%|58.8396717s|58.3054002s|
|Pow  |0.0565113567%|51.8469237s|51.8176243s|
|Round|  3.87328635%|4.83688612s|4.64953967s|
|Sin  |  1.90053224%|51.5278393s|50.5485361s|
|Tan  | 0.441676743%|58.7293771s|58.4699831s|

FYI. @AndyAyersMS 

Let me know what else, if anything, is required to get this merged.