@AndyAyersMS 2016-05-23
LGTM.

Something for future follow-up: try and run Linux on same HW as Windows, and compare relative perf (eg dual-boot a machine or similar). Data above and [some experiments I've done](https://github.com/AndyAyersMS/PerformanceExplorer/issues/4) suggest Linux math function performance may not be very good.

Also windows 10 math libs seem to light up on AVX2 capable machines. So for Windows, you might want to measure on both older and newer HW.