@tannergooding 2016-05-24
@AndyAyersMS, I did some investigations on Linux and it seems that the slowness is coming from the libm implementation being linked in (`usr/lib/x86_64-linux-gnu/libm.a`).

I tried swapping the implementation out with `amdlibm.a` and saw the numbers drop to be on par with Windows.