@AndyAyersMS 2016-05-24
Interesting... so do you think we should be linking amdlibm.a instead? Seems like we should open an issue for this.